my_blog/requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.10.7
django-allauth==0.32.0
django-ckeditor==5.2.2
django-classy-tags==0.8.0
django-cleanup==0.4.2
django-crispy-forms==1.6.1
django-hitcount==1.2.2
django-taggit==0.22.1
django-taggit-templatetags2==1.6.1
django-templatetag-sugar==1.0
djangorestframework==3.6.3
gunicorn==19.7.1
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.1
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
six==1.10.0
whitenoise==3.3.0
